pkg/bkop: include super_read_only value in ServerStatus
